next steps

+++ postgresql_crud/postgresql_crud_common/postgresql_crud_types_macro_logic_reuse/src/lib.rs
@@ -3484,7 +3484,7 @@ impl ShouldWhereElementFieldsBePublic<'_> {
         }
     }
 }
-fn generate_serde_deserialize_double_quotes_token_stream(postgresql_type_ident_where_element_tokens_upper_camel_case: &dyn naming::StdFmtDisplayPlusQuoteToTokens, length: std::primitive::u64) -> (
+fn generate_serde_deserialize_double_quotes_token_stream(postgresql_type_ident_where_element_tokens_upper_camel_case: &dyn naming::StdFmtDisplayPlusQuoteToTokens, length: std::primitive::usize) -> (
     proc_macro2::TokenStream,
     proc_macro2::TokenStream,
     proc_macro2::TokenStream
